ref(protocol): Rename session module for consistency
